# api/v1/views/documention/tasks_api_docs.yml
openapi: 3.0.0
info:
  title: Task Manager API
  description: API for managing tasks.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1

paths:
  /tasks:
    get:
      summary: Retrieve all tasks
      description: Retrieves a list of all task objects.
      responses:
        '200':
          description: A list of task objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      description: Creates a new task object.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  '/tasks/{task_id}':
    get:
      summary: Retrieve a task by ID
      description: Retrieves a specific task object by its ID.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task object retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    put:
      summary: Update a task by ID
      description: Updates a specific task object by its ID.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task updated successfully
        '400':
          description: Invalid request data
        '404':
          description: Task not found

    delete:
      summary: Delete a task by ID
      description: Deletes a specific task object by its ID.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_name:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - task_name
        - start_date
        - end_date